added user restriction for the user

+++ sales.php
@@ -36,7 +36,7 @@ $sales = find_all_sale();
                 <th class="text-center" style="width: 10%;"> Quantity </th>
                 <th class="text-center" style="width: 10%;"> Total </th>
                 <th class="text-center" style="width: 10%;"> Price </th>
-                <th class="text-center" style="width: 10%;"> Sell Date </th>
+                <th class="text-center" style="width: 10%;"> Date </th>
                 <th class="text-center" style="width: 10%;"> Product Added </th>
                 <th class="text-center" style="width: 10%;"> Date Purchased </th>
                 <th class="text-center" style="width: 10%;"> Supplier</th>
@@ -53,46 +53,48 @@ $sales = find_all_sale();
                <td><?php echo remove_junk($sale['category']); ?></td>
                <td><?php echo remove_junk($sale['sizes']); ?></td>
                <td class="text-center"><?php echo (int)$sale['qty']; ?></td>
-                <td class="text-center"><?php echo number_format((float)($sale['price'] * $sale['qty']), 2, '.', ','); ?></td>
+               <td class="text-center"><?php echo remove_junk($sale['price'] * $sale['qty']); ?></td>
                <td class="text-center"><?php echo remove_junk( $sale['price'] ); ?></td>
+               <td class="text-center"><?php echo remove_junk($sale['date']); ?></td>
                <td class="text-center"><?php echo remove_junk($sale['date']); ?></td>
                <td class="text-center"><?php echo remove_junk($sale['product_added']); ?></td>
                <td class="text-center"><?php echo remove_junk($sale['date_purchased']); ?></td>
                <td class="text-center"><?php echo remove_junk($sale['supplier']); ?></td>
                <td class="text-center"><?php echo remove_junk($sale['or_number']); ?></td>
-               <td class="text-center"><?php echo $sale['remarks']; ?></td>
-               <td class="text-center">
-                  <div class="btn-group">
-                     <a href="edit_sale.php?id=<?php echo (int)$sale['id'];?>" class="btn btn-warning btn-xs"  title="Edit" data-toggle="tooltip">
-                       <span class="glyphicon glyphicon-edit"></span>
-                     </a>
-                     
-                    <button type="button" class="btn btn-danger btn-xs" title="Delete" data-toggle="modal" data-target="#deleteModal<?php echo (int)$sale['id']; ?>">
-                      <span class="glyphicon glyphicon-trash"></span>
-                    </button>
+              <?php if($user['user_level'] == 1): ?>
+                       <td class="text-center">
+                        <div class="btn-group">
+                         <a href="edit_sale.php?id=<?php echo (int)$sale['id'];?>" class="btn btn-warning btn-xs"  title="Edit" data-toggle="tooltip">
+                           <span class="glyphicon glyphicon-edit"></span>
+                         </a>
+                         
+                        <button type="button" class="btn btn-danger btn-xs" title="Delete" data-toggle="modal" data-target="#deleteModal<?php echo (int)$sale['id']; ?>">
+                          <span class="glyphicon glyphicon-trash"></span>
+                        </button>
 
-                    <!-- Delete Confirmation Modal -->
-                    <div class="modal fade" id="deleteModal<?php echo (int)$sale['id']; ?>" tabindex="-1" role="dialog" aria-labelledby="deleteModalLabel<?php echo (int)$sale['id']; ?>" aria-hidden="true">
-                      <div class="modal-dialog" role="document">
-                        <div class="modal-content">
-                          <div class="modal-header">
+                        <!-- Delete Confirmation Modal -->
+                        <div class="modal fade" id="deleteModal<?php echo (int)$sale['id']; ?>" tabindex="-1" role="dialog" aria-labelledby="deleteModalLabel<?php echo (int)$sale['id']; ?>" aria-hidden="true">
+                          <div class="modal-dialog" role="document">
+                          <div class="modal-content">
+                            <div class="modal-header">
                             <h4 class="modal-title" id="deleteModalLabel<?php echo (int)$sale['id']; ?>">Confirm Deletion</h4>
                             <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                               <span aria-hidden="true">&times;</span>
                             </button>
-                          </div>
-                          <div class="modal-body">
+                            </div>
+                            <div class="modal-body">
                             Are you sure you want to delete this sale?
-                          </div>
-                          <div class="modal-footer">
+                            </div>
+                            <div class="modal-footer">
                             <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
                             <a href="delete_sale.php?id=<?php echo (int)$sale['id'];?>" class="btn btn-danger">Delete</a>
+                            </div>
+                          </div>
                           </div>
                         </div>
-                      </div>
-                    </div>
-                  </div>
-               </td>
+                        </div>
+                       </td>
+              <?php endif; ?>
              </tr>
              <?php endforeach;?>
            </tbody>
